What about automatic check access
This feature provides the hability to automatic check if the user can access or not the action, based on a map of settings to action.

So if it's set defaultMap as true, it gonna check what is the action been access and map that as an action of Authorizit, and get the modelClass as the resource to use.